引入nginx服务器,修改nginx.conf配置文件,解决ajax跨越请求问题
