+++ blog-web/server.js
@@ -53,4 +53,4 @@ http.createServer(function(req, res) {
 
 }).listen(8000, "127.0.0.1");
 
-console.log("Server running at http://127.0.0.1/");
+console.log("Server running at http://127.0.0.1:8000/");
