+++ controller-admin-1001/src/main/java/com/eva/ControllerAdminApplication_1001.java
@@ -1,10 +1,8 @@
 package com.eva;
 
 
-import com.netflix.hystrix.contrib.javanica.annotation.HystrixCommand;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 import org.springframework.cloud.netflix.hystrix.dashboard.EnableHystrixDashboard;
 import org.springframework.cloud.openfeign.EnableFeignClients;
@@ -17,10 +15,10 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 @EnableEurekaClient //开启Eureka
 @EnableFeignClients
 @EnableHystrixDashboard //开启
-public class ControllerAdminApplication
+public class ControllerAdminApplication_1001
 {
     public static void main( String[] args )
     {
-        SpringApplication.run(ControllerAdminApplication.class,args);
+        SpringApplication.run(ControllerAdminApplication_1001.class,args);
     }
 }
