+++ nginx-1.19.3/contrib/README
@@ -1,0 +1,21 @@
+
+geo2nginx.pl 		by Andrei Nigmatulin
+
+	The perl script to convert CSV geoip database ( free download
+	at http://www.maxmind.com/app/geoip_country ) to format, suitable
+	for use by the ngx_http_geo_module.
+
+
+unicode2nginx		by Maxim Dounin
+
+	The perl script to convert unicode mappings ( available
+	at http://www.unicode.org/Public/MAPPINGS/ ) to the nginx
+	configuration file format.
+	Two generated full maps for windows-1251 and koi8-r.
+
+
+vim			by Evan Miller
+
+	Syntax highlighting of nginx configuration for vim, to be
+	placed into ~/.vim/.
+
