+++ nginx-1.19.3/contrib/unicode2nginx/win-utf
@@ -1,0 +1,130 @@
+charset_map  windows-1251  utf-8 {
+
+    80  D082 ; #CYRILLIC CAPITAL LETTER DJE
+    81  D083 ; #CYRILLIC CAPITAL LETTER GJE
+    82  E2809A ; #SINGLE LOW-9 QUOTATION MARK
+    83  D193 ; #CYRILLIC SMALL LETTER GJE
+    84  E2809E ; #DOUBLE LOW-9 QUOTATION MARK
+    85  E280A6 ; #HORIZONTAL ELLIPSIS
+    86  E280A0 ; #DAGGER
+    87  E280A1 ; #DOUBLE DAGGER
+    88  E282AC ; #EURO SIGN
+    89  E280B0 ; #PER MILLE SIGN
+    8A  D089 ; #CYRILLIC CAPITAL LETTER LJE
+    8B  E280B9 ; #SINGLE LEFT-POINTING ANGLE QUOTATION MARK
+    8C  D08A ; #CYRILLIC CAPITAL LETTER NJE
+    8D  D08C ; #CYRILLIC CAPITAL LETTER KJE
+    8E  D08B ; #CYRILLIC CAPITAL LETTER TSHE
+    8F  D08F ; #CYRILLIC CAPITAL LETTER DZHE
+    90  D192 ; #CYRILLIC SMALL LETTER DJE
+    91  E28098 ; #LEFT SINGLE QUOTATION MARK
+    92  E28099 ; #RIGHT SINGLE QUOTATION MARK
+    93  E2809C ; #LEFT DOUBLE QUOTATION MARK
+    94  E2809D ; #RIGHT DOUBLE QUOTATION MARK
+    95  E280A2 ; #BULLET
+    96  E28093 ; #EN DASH
+    97  E28094 ; #EM DASH
+    99  E284A2 ; #TRADE MARK SIGN
+    9A  D199 ; #CYRILLIC SMALL LETTER LJE
+    9B  E280BA ; #SINGLE RIGHT-POINTING ANGLE QUOTATION MARK
+    9C  D19A ; #CYRILLIC SMALL LETTER NJE
+    9D  D19C ; #CYRILLIC SMALL LETTER KJE
+    9E  D19B ; #CYRILLIC SMALL LETTER TSHE
+    9F  D19F ; #CYRILLIC SMALL LETTER DZHE
+    A0  C2A0 ; #NO-BREAK SPACE
+    A1  D08E ; #CYRILLIC CAPITAL LETTER SHORT U
+    A2  D19E ; #CYRILLIC SMALL LETTER SHORT U
+    A3  D088 ; #CYRILLIC CAPITAL LETTER JE
+    A4  C2A4 ; #CURRENCY SIGN
+    A5  D290 ; #CYRILLIC CAPITAL LETTER GHE WITH UPTURN
+    A6  C2A6 ; #BROKEN BAR
+    A7  C2A7 ; #SECTION SIGN
+    A8  D081 ; #CYRILLIC CAPITAL LETTER IO
+    A9  C2A9 ; #COPYRIGHT SIGN
+    AA  D084 ; #CYRILLIC CAPITAL LETTER UKRAINIAN IE
+    AB  C2AB ; #LEFT-POINTING DOUBLE ANGLE QUOTATION MARK
+    AC  C2AC ; #NOT SIGN
+    AD  C2AD ; #SOFT HYPHEN
+    AE  C2AE ; #REGISTERED SIGN
+    AF  D087 ; #CYRILLIC CAPITAL LETTER YI
+    B0  C2B0 ; #DEGREE SIGN
+    B1  C2B1 ; #PLUS-MINUS SIGN
+    B2  D086 ; #CYRILLIC CAPITAL LETTER BYELORUSSIAN-UKRAINIAN I
+    B3  D196 ; #CYRILLIC SMALL LETTER BYELORUSSIAN-UKRAINIAN I
+    B4  D291 ; #CYRILLIC SMALL LETTER GHE WITH UPTURN
+    B5  C2B5 ; #MICRO SIGN
+    B6  C2B6 ; #PILCROW SIGN
+    B7  C2B7 ; #MIDDLE DOT
+    B8  D191 ; #CYRILLIC SMALL LETTER IO
+    B9  E28496 ; #NUMERO SIGN
+    BA  D194 ; #CYRILLIC SMALL LETTER UKRAINIAN IE
+    BB  C2BB ; #RIGHT-POINTING DOUBLE ANGLE QUOTATION MARK
+    BC  D198 ; #CYRILLIC SMALL LETTER JE
+    BD  D085 ; #CYRILLIC CAPITAL LETTER DZE
+    BE  D195 ; #CYRILLIC SMALL LETTER DZE
+    BF  D197 ; #CYRILLIC SMALL LETTER YI
+    C0  D090 ; #CYRILLIC CAPITAL LETTER A
+    C1  D091 ; #CYRILLIC CAPITAL LETTER BE
+    C2  D092 ; #CYRILLIC CAPITAL LETTER VE
+    C3  D093 ; #CYRILLIC CAPITAL LETTER GHE
+    C4  D094 ; #CYRILLIC CAPITAL LETTER DE
+    C5  D095 ; #CYRILLIC CAPITAL LETTER IE
+    C6  D096 ; #CYRILLIC CAPITAL LETTER ZHE
+    C7  D097 ; #CYRILLIC CAPITAL LETTER ZE
+    C8  D098 ; #CYRILLIC CAPITAL LETTER I
+    C9  D099 ; #CYRILLIC CAPITAL LETTER SHORT I
+    CA  D09A ; #CYRILLIC CAPITAL LETTER KA
+    CB  D09B ; #CYRILLIC CAPITAL LETTER EL
+    CC  D09C ; #CYRILLIC CAPITAL LETTER EM
+    CD  D09D ; #CYRILLIC CAPITAL LETTER EN
+    CE  D09E ; #CYRILLIC CAPITAL LETTER O
+    CF  D09F ; #CYRILLIC CAPITAL LETTER PE
+    D0  D0A0 ; #CYRILLIC CAPITAL LETTER ER
+    D1  D0A1 ; #CYRILLIC CAPITAL LETTER ES
+    D2  D0A2 ; #CYRILLIC CAPITAL LETTER TE
+    D3  D0A3 ; #CYRILLIC CAPITAL LETTER U
+    D4  D0A4 ; #CYRILLIC CAPITAL LETTER EF
+    D5  D0A5 ; #CYRILLIC CAPITAL LETTER HA
+    D6  D0A6 ; #CYRILLIC CAPITAL LETTER TSE
+    D7  D0A7 ; #CYRILLIC CAPITAL LETTER CHE
+    D8  D0A8 ; #CYRILLIC CAPITAL LETTER SHA
+    D9  D0A9 ; #CYRILLIC CAPITAL LETTER SHCHA
+    DA  D0AA ; #CYRILLIC CAPITAL LETTER HARD SIGN
+    DB  D0AB ; #CYRILLIC CAPITAL LETTER YERU
+    DC  D0AC ; #CYRILLIC CAPITAL LETTER SOFT SIGN
+    DD  D0AD ; #CYRILLIC CAPITAL LETTER E
+    DE  D0AE ; #CYRILLIC CAPITAL LETTER YU
+    DF  D0AF ; #CYRILLIC CAPITAL LETTER YA
+    E0  D0B0 ; #CYRILLIC SMALL LETTER A
+    E1  D0B1 ; #CYRILLIC SMALL LETTER BE
+    E2  D0B2 ; #CYRILLIC SMALL LETTER VE
+    E3  D0B3 ; #CYRILLIC SMALL LETTER GHE
+    E4  D0B4 ; #CYRILLIC SMALL LETTER DE
+    E5  D0B5 ; #CYRILLIC SMALL LETTER IE
+    E6  D0B6 ; #CYRILLIC SMALL LETTER ZHE
+    E7  D0B7 ; #CYRILLIC SMALL LETTER ZE
+    E8  D0B8 ; #CYRILLIC SMALL LETTER I
+    E9  D0B9 ; #CYRILLIC SMALL LETTER SHORT I
+    EA  D0BA ; #CYRILLIC SMALL LETTER KA
+    EB  D0BB ; #CYRILLIC SMALL LETTER EL
+    EC  D0BC ; #CYRILLIC SMALL LETTER EM
+    ED  D0BD ; #CYRILLIC SMALL LETTER EN
+    EE  D0BE ; #CYRILLIC SMALL LETTER O
+    EF  D0BF ; #CYRILLIC SMALL LETTER PE
+    F0  D180 ; #CYRILLIC SMALL LETTER ER
+    F1  D181 ; #CYRILLIC SMALL LETTER ES
+    F2  D182 ; #CYRILLIC SMALL LETTER TE
+    F3  D183 ; #CYRILLIC SMALL LETTER U
+    F4  D184 ; #CYRILLIC SMALL LETTER EF
+    F5  D185 ; #CYRILLIC SMALL LETTER HA
+    F6  D186 ; #CYRILLIC SMALL LETTER TSE
+    F7  D187 ; #CYRILLIC SMALL LETTER CHE
+    F8  D188 ; #CYRILLIC SMALL LETTER SHA
+    F9  D189 ; #CYRILLIC SMALL LETTER SHCHA
+    FA  D18A ; #CYRILLIC SMALL LETTER HARD SIGN
+    FB  D18B ; #CYRILLIC SMALL LETTER YERU
+    FC  D18C ; #CYRILLIC SMALL LETTER SOFT SIGN
+    FD  D18D ; #CYRILLIC SMALL LETTER E
+    FE  D18E ; #CYRILLIC SMALL LETTER YU
+    FF  D18F ; #CYRILLIC SMALL LETTER YA
+}
