+++ nginx-1.19.3/contrib/vim/ftplugin/nginx.vim
@@ -1,0 +1,1 @@
+setlocal commentstring=#\ %s
