+++ nginx-1.19.3/logs/nginx.pid
@@ -1,0 +1,1 @@
+17544
