+++ 接口地址.txt
@@ -1,0 +1,5 @@
+domain: 127.0.0.1
+
+sprt: 1000
+
+管理员接口 127.0.0.1:1000/blog/ad/*
